refactor: move to a daemon-focused backend
